+++ provider/datagen/src/transform/cldr/decimal/symbols.rs
@@ -1,0 +1,133 @@
+use crate::transform::cldr::cldr_serde;
+use icu_decimal::provider::*;
+use icu_locid::extensions::unicode::Value;
+use icu_locid::extensions_unicode_key as key;
+use icu_provider::datagen::IterableDataProvider;
+use icu_provider::prelude::*;
+use std::borrow::Cow;
+use std::convert::TryFrom;
+use tinystr::TinyAsciiStr;
+
+impl DataProvider<DecimalSymbolsV1Marker> for crate::DatagenProvider {
+    fn load(&self, req: DataRequest) -> Result<DataResponse<DecimalSymbolsV1Marker>, DataError> {
+        let langid = req.locale.get_langid();
+
+        let resource: &cldr_serde::numbers::Resource = self
+            .source
+            .cldr()?
+            .numbers()
+            .read_and_parse(&langid, "numbers.json")?;
+
+        let numbers = &resource
+            .main
+            .0
+            .get(&langid)
+            .expect("CLDR file contains the expected language")
+            .numbers;
+
+        let nsname = match req.locale.get_unicode_ext(&key!("nu")) {
+            Some(v) => *v
+                .as_tinystr_slice()
+                .first()
+                .expect("expecting subtag if key is present"),
+            None => numbers.default_numbering_system,
+        };
+
+        let mut result =
+            DecimalSymbolsV1::try_from(NumbersWithNumsys(numbers, nsname)).map_err(|s| {
+                DataError::custom("Could not create decimal symbols")
+                    .with_display_context(&s)
+                    .with_display_context(&nsname)
+            })?;
+
+        result.digits = self.get_digits_for_numbering_system(nsname)?;
+
+        Ok(DataResponse {
+            metadata: Default::default(),
+            payload: Some(DataPayload::from_owned(result)),
+        })
+    }
+}
+
+impl IterableDataProvider<DecimalSymbolsV1Marker> for crate::DatagenProvider {
+    fn supported_locales(&self) -> Result<Vec<DataLocale>, DataError> {
+        Ok(self
+            .source
+            .cldr()?
+            .numbers()
+            .list_langs()?
+            .flat_map(|langid| {
+                let last = DataLocale::from(&langid);
+                self.get_supported_numsys_for_langid_without_default(&langid)
+                    .expect("All languages from list_langs should be present")
+                    .into_iter()
+                    .map(move |nsname| {
+                        let mut data_locale = DataLocale::from(&langid);
+                        data_locale.set_unicode_ext(
+                            key!("nu"),
+                            Value::try_from_single_subtag(nsname.as_bytes())
+                                .expect("CLDR should have valid numbering system names"),
+                        );
+                        data_locale
+                    })
+                    .chain(core::iter::once(last))
+            })
+            .collect())
+    }
+}
+
+struct NumbersWithNumsys<'a>(pub &'a cldr_serde::numbers::Numbers, pub TinyAsciiStr<8>);
+
+impl TryFrom<NumbersWithNumsys<'_>> for DecimalSymbolsV1<'static> {
+    type Error = Cow<'static, str>;
+
+    fn try_from(other: NumbersWithNumsys<'_>) -> Result<Self, Self::Error> {
+        let NumbersWithNumsys(numbers, nsname) = other;
+        let symbols = numbers
+            .numsys_data
+            .symbols
+            .get(&nsname)
+            .ok_or("Could not find symbols for numbering system")?;
+        let formats = numbers
+            .numsys_data
+            .formats
+            .get(&nsname)
+            .ok_or("Could not find formats for numbering system")?;
+        let parsed_pattern: super::decimal_pattern::DecimalPattern = formats
+            .standard
+            .parse()
+            .map_err(|s: super::decimal_pattern::Error| s.to_string())?;
+
+        Ok(Self {
+            minus_sign_affixes: parsed_pattern.localize_sign(&symbols.minus_sign),
+            plus_sign_affixes: parsed_pattern.localize_sign(&symbols.plus_sign),
+            decimal_separator: Cow::Owned(symbols.decimal.clone()),
+            grouping_separator: Cow::Owned(symbols.group.clone()),
+            grouping_sizes: GroupingSizesV1 {
+                primary: parsed_pattern.positive.primary_grouping,
+                secondary: parsed_pattern.positive.secondary_grouping,
+                min_grouping: numbers.minimum_grouping_digits,
+            },
+            digits: Default::default(), // to be filled in
+        })
+    }
+}
+
+#[test]
+fn test_basic() {
+    use icu_locid::locale;
+
+    let provider = crate::DatagenProvider::for_test();
+
+    let ar_decimal: DataPayload<DecimalSymbolsV1Marker> = provider
+        .load(DataRequest {
+            locale: &locale!("ar-EG").into(),
+            metadata: Default::default(),
+        })
+        .unwrap()
+        .take_payload()
+        .unwrap();
+
+    assert_eq!(ar_decimal.get().decimal_separator, "٫");
+    assert_eq!(ar_decimal.get().digits[0], '٠');
+}
